+++ Framework-Headers/Frameworks/MediaPlayer._framework/Objc/MPAVItemTimeMarkerEvent.h
@@ -24,7 +24,7 @@
 @property (strong, nonatomic) MPTimeMarker *closedCaptionMarker; // @synthesize closedCaptionMarker=_closedCaptionMarker;
 
 - (void).cxx_destruct;
-- (id)crossedMarkerOfType:(int)arg1;
+- (id)crossedMarkerOfType:(long long)arg1;
 - (id)initWithTimeMarkerNotificationUserInfo:(id)arg1;
 
 @end
