+++ Framework-Headers/Frameworks/UIKit._framework/Objc/UIInputSwitcherSelectionExtraView.h
@@ -12,7 +12,7 @@ __attribute__((visibility("hidden")))
 @interface UIInputSwitcherSelectionExtraView : UIView
 {
     double m_pointerOffset;
-    int _roundedCorners;
+    unsigned long long _roundedCorners;
     UIKeyboardMenuView *_menu;
     struct CGRect _keyRect;
 }
@@ -20,7 +20,7 @@ __attribute__((visibility("hidden")))
 @property (nonatomic) struct CGRect keyRect; // @synthesize keyRect=_keyRect;
 @property (nonatomic) UIKeyboardMenuView *menu; // @synthesize menu=_menu;
 @property (nonatomic) double pointerOffset; // @synthesize pointerOffset=m_pointerOffset;
-@property (nonatomic) int roundedCorners; // @synthesize roundedCorners=_roundedCorners;
+@property (nonatomic) unsigned long long roundedCorners; // @synthesize roundedCorners=_roundedCorners;
 
 - (void)drawRect:(struct CGRect)arg1;
 - (id)initWithFrame:(struct CGRect)arg1;
