+++ Framework-Headers/Frameworks/UIKit._framework/Objc/UIKBBackdropView.h
@@ -16,14 +16,14 @@
 @property (strong, nonatomic) UIImage *imageForCorners; // @synthesize imageForCorners=_imageForCorners;
 
 - (BOOL)_applyCornerMaskToSelf;
-- (id)_generateCornerContentsImage:(int)arg1;
+- (id)_generateCornerContentsImage:(unsigned long long)arg1;
 - (void)_setRenderConfig:(id)arg1;
 - (void)dealloc;
 - (id)initWithFrame:(struct CGRect)arg1 style:(long long)arg2;
 - (void)setFrame:(struct CGRect)arg1;
 - (int)textEffectsVisibilityLevel;
 - (void)transitionToStyle:(long long)arg1;
-- (void)updateFrame:(struct CGRect)arg1 withCorners:(int)arg2;
+- (void)updateFrame:(struct CGRect)arg1 withCorners:(unsigned long long)arg2;
 
 @end
