+++ Framework-Headers/Frameworks/UIKit._framework/Objc/UIKBSplitTraits.h
@@ -8,16 +8,16 @@
 
 @interface UIKBSplitTraits : NSObject
 {
-    int _corners;
+    unsigned long long _corners;
     struct CGRect _leftFrame;
     struct CGRect _rightFrame;
 }
 
-@property (readonly, nonatomic) int corners; // @synthesize corners=_corners;
+@property (readonly, nonatomic) unsigned long long corners; // @synthesize corners=_corners;
 @property (readonly, nonatomic) struct CGRect leftFrame; // @synthesize leftFrame=_leftFrame;
 @property (readonly, nonatomic) struct CGRect rightFrame; // @synthesize rightFrame=_rightFrame;
 
-+ (id)traitsWithLeftFrame:(struct CGRect)arg1 rightFrame:(struct CGRect)arg2 corners:(int)arg3;
++ (id)traitsWithLeftFrame:(struct CGRect)arg1 rightFrame:(struct CGRect)arg2 corners:(unsigned long long)arg3;
 - (id)initWithLeftFrame:(struct CGRect)arg1 rightFrame:(struct CGRect)arg2 corners:(unsigned long long)arg3;
 
 @end
