+++ Framework-Headers/PrivateFrameworks/CoreSuggestionsInternals._framework/Objc/SGDatabaseMigratorEntities.h
@@ -11,7 +11,7 @@
 }
 
 + (id)migrations;
-+ (id)skipFromZeroSchema:(unsigned long long *)arg1;
++ (id)skipFromZeroSchema:(unsigned int *)arg1;
 
 @end
