+++ Framework-Headers/PrivateFrameworks/CoreSuggestionsInternals._framework/Objc/SGStorageLocation.h
@@ -14,7 +14,7 @@
 
 @interface SGStorageLocation : NSObject <NSCopying, NSSecureCoding, SGEventLocationForGeocode>
 {
-    unsigned long long _locationType;
+    int _locationType;
     NSString *_label;
     NSString *_address;
     NSString *_airportCode;
@@ -34,7 +34,7 @@
 @property (readonly) unsigned long long hash;
 @property (readonly, nonatomic) NSString *label; // @synthesize label=_label;
 @property (readonly, nonatomic) double latitude; // @synthesize latitude=_latitude;
-@property (readonly, nonatomic) unsigned long long locationType; // @synthesize locationType=_locationType;
+@property (readonly, nonatomic) int locationType; // @synthesize locationType=_locationType;
 @property (readonly, nonatomic) double longitude; // @synthesize longitude=_longitude;
 @property (readonly, nonatomic) double quality; // @synthesize quality=_quality;
 @property (readonly) Class superclass;
@@ -57,9 +57,9 @@
 - (id)geocodedLocationWithLabel:(id)arg1 address:(id)arg2 latitude:(double)arg3 longitude:(double)arg4 accuracy:(double)arg5 handle:(id)arg6;
 - (id)geocodedLocationWithLatitude:(double)arg1 longitude:(double)arg2 accuracy:(double)arg3 handle:(id)arg4;
 - (id)initWithCoder:(id)arg1;
-- (id)initWithType:(unsigned long long)arg1 label:(id)arg2 address:(id)arg3 airportCode:(id)arg4 accuracy:(double)arg5 quality:(double)arg6;
-- (id)initWithType:(unsigned long long)arg1 label:(id)arg2 address:(id)arg3 airportCode:(id)arg4 latitude:(double)arg5 longitude:(double)arg6 accuracy:(double)arg7 quality:(double)arg8 handle:(id)arg9;
-- (id)initWithType:(unsigned long long)arg1 label:(id)arg2 airportCode:(id)arg3 latitude:(double)arg4 longitude:(double)arg5 accuracy:(double)arg6 quality:(double)arg7;
+- (id)initWithType:(int)arg1 label:(id)arg2 address:(id)arg3 airportCode:(id)arg4 accuracy:(double)arg5 quality:(double)arg6;
+- (id)initWithType:(int)arg1 label:(id)arg2 address:(id)arg3 airportCode:(id)arg4 latitude:(double)arg5 longitude:(double)arg6 accuracy:(double)arg7 quality:(double)arg8 handle:(id)arg9;
+- (id)initWithType:(int)arg1 label:(id)arg2 airportCode:(id)arg3 latitude:(double)arg4 longitude:(double)arg5 accuracy:(double)arg6 quality:(double)arg7;
 - (BOOL)isEqual:(id)arg1;
 - (BOOL)isEqualToStorageLocation:(id)arg1;
