+++ Framework-Headers/PrivateFrameworks/Message._framework/Objc/MFBaseMessage-Protocol.h
@@ -16,9 +16,9 @@
 @property (readonly, nonatomic) BOOL deleted;
 @property (readonly, nonatomic) BOOL flagged;
 @property (readonly, nonatomic, getter=isKnownToHaveAttachments) BOOL knownToHaveAttachments;
-@property (readonly, nonatomic) unsigned int libraryID;
+@property (readonly, nonatomic) long long libraryID;
 @property (readonly, nonatomic, getter=isLibraryMessage) BOOL libraryMessage;
-@property (readonly, nonatomic) unsigned int mailboxID;
+@property (readonly, nonatomic) long long mailboxID;
 @property (readonly, nonatomic) long long messageIDHash;
 @property (readonly, nonatomic) BOOL read;
 @property (readonly, nonatomic) BOOL senderVIP;
